add more coverage tests

+++ contracts/SingularityOracle.sol
@@ -10,9 +10,9 @@ import './interfaces/ISingularityOracle.sol';
  */
 contract SingularityOracle is ISingularityOracle {
 	struct PriceData {
-		uint price;
-		uint updatedAt;
-		uint nonce;
+		uint256 price;
+		uint256 updatedAt;
+		uint256 nonce;
 	}
 
 	address admin;
@@ -27,29 +27,29 @@ contract SingularityOracle is ISingularityOracle {
 
 	/// @dev Does NOT validate returned answers
 	/// @dev Need to do data validation in interacting contract
-	function getLatestRound(address _token) public view override returns (uint price, uint updatedAt) {
+	function getLatestRound(address _token) public view override returns (uint256 price, uint256 updatedAt) {
 		PriceData[] memory prices = allPrices[_token];
 		price = prices[prices.length - 1].price;
 		updatedAt = prices[prices.length - 1].updatedAt;
 	}
 
-	function getLatestRounds(address[] calldata tokens) external view returns (uint[] memory prices, uint[] memory updatedAts) {
-		for (uint i; i < tokens.length; i++) {
-			(uint price, uint updatedAt) = getLatestRound(tokens[i]);
+	function getLatestRounds(address[] calldata tokens) external view returns (uint256[] memory prices, uint256[] memory updatedAts) {
+		for (uint256 i; i < tokens.length; i++) {
+			(uint256 price, uint256 updatedAt) = getLatestRound(tokens[i]);
 			prices[i] = price;
 			updatedAts[i] = updatedAt;
 		}
 	}
 
-	function pushPrice(address _token, uint _price) public {
+	function pushPrice(address _token, uint256 _price) public {
 		require(pushers[msg.sender], "SingularityOracle: NOT_PUSHER");
 		PriceData[] storage prices = allPrices[_token];
 		prices.push(PriceData({ price: _price, updatedAt: block.timestamp, nonce: prices.length }));
 	}
 
-	function pushPrices(address[] calldata tokens, uint[] calldata prices) external {
+	function pushPrices(address[] calldata tokens, uint256[] calldata prices) external {
 		require(tokens.length == prices.length, "SingularityOracle: NOT_SAME_LENGTH");
-		for(uint i; i < tokens.length; i++) {
+		for(uint256 i; i < tokens.length; i++) {
 			pushPrice(tokens[i], prices[i]);
 		}
 	}
