+++ contracts/interfaces/IWETH.sol
@@ -4,5 +4,5 @@ pragma solidity ^0.8.11;
 
 interface IWETH {
     function deposit() external payable;
-    function withdraw(uint) external;
+    function withdraw(uint256) external;
 }
