+++ contracts/utils/ReentrancyGuard.sol
@@ -5,7 +5,7 @@ pragma solidity >=0.8.0;
 /// @author Solmate (https://github.com/Rari-Capital/solmate/blob/main/src/utils/ReentrancyGuard.sol)
 /// @author Modified from OpenZeppelin (https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/contracts/security/ReentrancyGuard.sol)
 abstract contract ReentrancyGuard {
-    uint private locked = 1;
+    uint256 private locked = 1;
 
     modifier nonReentrant() {
         require(locked == 1, "REENTRANCY");
